# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shubham Motors AI Voice Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: score
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a single lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/import:
    post:
      operationId: importLeads
      tags: [leads]
      summary: Bulk import leads from CSV/JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/call:
    post:
      operationId: triggerOutboundCall
      tags: [leads]
      summary: Trigger an outbound call to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallInitiated"

  # ── CALLS ──────────────────────────────────────────────────────────────────
...
  /webhooks/exotel/inbound:
    post:
      operationId: exotelInboundWebhook
      tags: [webhooks]
      summary: Handle inbound call from Exotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExotelWebhookPayload"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"
  /webhooks/exotel/status:
    post:
      operationId: exotelStatusWebhook
      tags: [webhooks]
      summary: Handle call status updates from Exotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExotelWebhookPayload"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"
  /webhooks/voice/stream:
    post:
      operationId: voiceStreamWebhook
      tags: [webhooks]
      summary: Handle real-time voice audio stream (Sarvam STT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceStreamPayload"
      responses:
        "200":
          description: TTS audio response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceStreamResponse"

  # ── KNOWLEDGE BASE ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, phone, status, score, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, interested, hot, converted, lost, followup]
        score:
          type: integer
          description: Lead score 0-100
        interestedModel:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        lastCallId:
          type: ["integer", "null"]
        nextFollowupAt:
          type: ["string", "null"]
        intentSummary:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        interestedModel:
          type: string
        source:
          type: string
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
        score:
          type: integer
        interestedModel:
          type: string
        language:
          type: string
        notes:
          type: string
        nextFollowupAt:
          type: string
        intentSummary:
          type: string
    LeadImport:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadInput"
    ImportResult:
      type: object
      required: [imported, failed, total]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: string

    # ── CALL ────────────────────────────────────────────────────────────────
...
    CallInitiated:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        callSid:
          type: ["string", "null"]
...
    ExotelWebhookPayload:
      type: object
      properties:
        CallSid:
          type: string
        From:
          type: string
        To:
          type: string
        Status:
          type: string
        RecordingUrl:
          type: ["string", "null"]
        Duration:
          type: ["string", "null"]
        Direction:
          type: ["string", "null"]
    VoiceStreamPayload:
      type: object
      required: [callSid, audioData, language]
      properties:
        callSid:
          type: string
        audioData:
          type: string
          description: Base64 encoded audio
        language:
          type: string
    VoiceStreamResponse:
      type: object
      required: [text, audioData]
      properties:
        text:
          type: string
        audioData:
          type: string
          description: Base64 encoded TTS audio
        shouldTransfer:
          type: boolean
        transferNumber:
          type: ["string", "null"]
    WebhookAck:
      type: object
      required: [received]
      properties:
        received:
          type: boolean

    # ── KNOWLEDGE ────────────────────────────────────────────────────────────
...