# lib/api-zod/src/generated/types/exotelWebhookPayload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Motors AI Voice Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ExotelWebhookPayload {
  CallSid?: string;
  From?: string;
  To?: string;
  Status?: string;
  /** @nullable */
  RecordingUrl?: string | null;
  /** @nullable */
  Duration?: string | null;
  /** @nullable */
  Direction?: string | null;
}