# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Motors AI Voice Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadUpdate {
  name?: string;
  phone?: string;
  email?: string;
  status?: string;
  score?: number;
  interestedModel?: string;
  language?: string;
  notes?: string;
  nextFollowupAt?: string;
  intentSummary?: string;
}