# lib/api-zod/src/generated/types/voiceStreamPayload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Motors AI Voice Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceStreamPayload {
  callSid: string;
  /** Base64 encoded audio */
  audioData: string;
  language: string;
}